requirements.txt
absl-py==0.11.0
appdirs==1.4.3
apturl==0.5.2
astunparse==1.6.3
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
dicttoxml==1.7.4
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.3.1
duplicity==0.8.12.0
entrypoints==0.3
Fabric3==1.14.post1
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
ovs==2.13.1
pandas==1.2.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
plotly==4.14.3
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
pyglet==1.5.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-distutils-extra==2.39
python-vagrant==0.5.15
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.6
scapy==2.4.4
scipy==1.5.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
sortedcontainers==2.1.0
ssh-import-id==5.10
systemd-python==234
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tflite-runtime==2.5.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.17
wadllib==1.3.3
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
